--- src/tranprobs.h
@@ -1,22 +1,0 @@
-#ifndef _unmarked_TRANPROBS_H
-#define _unmarked_TRANPROBS_H
-
-#include <RcppArmadillo.h>
-
-
-// constant model
-//void tp1(arma::mat& g3, int lk, double gam, double om);
-void tp1(arma::mat& g3, int nrI, int nrI1, Rcpp::IntegerVector N, arma::imat I, arma::imat I1, Rcpp::List Ib, Rcpp::List Ip, double gam, double om);
-
-
-
-// autoregressive model
-void tp2(arma::mat& g3, int lk, double gam, double om);
-
-
-
-// trend model
-void tp3(arma::mat& g3, int lk, double gam);
-
-
-#endif
